+++ WordPress/Tests/DB/DirectDatabaseQueryUnitTest.php
@@ -7,7 +7,7 @@
  * @license https://opensource.org/licenses/MIT MIT
  */
 
-namespace WordPress\Tests\VIP;
+namespace WordPress\Tests\DB;
 
 use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
 
@@ -18,8 +18,7 @@ use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
  *
  * @since   0.3.0
  * @since   0.13.0 Class name changed: this class is now namespaced.
- * @since   0.15.0 The sniff has been deprecated. This unit test file now
- *                 only tests that the deprecation warnings are correctly thrown.
+ * @since   0.15.0 This sniff has been moved from the `VIP` category to the `DB` category.
  */
 class DirectDatabaseQueryUnitTest extends AbstractSniffUnitTest {
 
@@ -29,7 +28,23 @@ class DirectDatabaseQueryUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of errors>
 	 */
 	public function getErrorList() {
-		return array();
+		return array(
+			6   => 1,
+			8   => 1,
+			32  => 1,
+			34  => 1,
+			50  => 1,
+			78  => 1,
+			79  => 1,
+			80  => 1,
+			170 => 1,
+			178 => 1,
+			190 => 1,
+			250 => 1,
+			257 => 1,
+			274 => 1,
+		);
+
 	}
 
 	/**
@@ -39,7 +54,12 @@ class DirectDatabaseQueryUnitTest extends AbstractSniffUnitTest {
 	 */
 	public function getWarningList() {
 		return array(
-			1 => 2,
+			6   => 1,
+			17  => 1,
+			38  => 1,
+			50  => 1,
+			112 => 1,
+			250 => 1,
 		);
 
 	}
