+++ WordPress/ruleset.xml
@@ -15,4 +15,28 @@
 			<property name="exclude" value="obfuscation"/>
 		</properties>
 	</rule>
+
+	<!--
+	#############################################################################
+	Account for deprecated sniffs.
+	These directives can be removed when the deprecated sniffs are removed.
+	#############################################################################
+	-->
+
+	<!-- Prevent deprecation notice when the sniff is not explicitely included. -->
+	<rule ref="WordPress.VIP.DirectDatabaseQuery.DeprecatedSniff">
+		<severity>0</severity>
+	</rule>
+
+	<!-- Prevent duplicate messages from deprecated sniff. -->
+	<rule ref="WordPress.VIP.DirectDatabaseQuery.SchemaChange">
+		<severity>0</severity>
+	</rule>
+	<rule ref="WordPress.VIP.DirectDatabaseQuery.DirectQuery">
+		<severity>0</severity>
+	</rule>
+	<rule ref="WordPress.VIP.DirectDatabaseQuery.NoCaching">
+		<severity>0</severity>
+	</rule>
+
 </ruleset>
